refactor: ContentPlaceRepository를 ContentRepository로 통합
