--- android/app/src/main/java/com/on/turip/data/content/place/datasource/ContentPlaceRemoteDataSource.kt
@@ -1,8 +1,0 @@
-package com.on.turip.data.content.place.datasource
-
-import com.on.turip.core.result.TuripResult
-import com.on.turip.data.content.place.dto.ContentPlacesResponse
-
-interface ContentPlaceRemoteDataSource {
-    suspend fun getTrip(contentId: Long): TuripResult<ContentPlacesResponse>
-}
