+++ android/app/src/main/java/com/on/turip/data/content/repository/DefaultContentRepository.kt
@@ -3,11 +3,13 @@ package com.on.turip.data.content.repository
 import com.on.turip.core.result.TuripResult
 import com.on.turip.core.result.mapCatching
 import com.on.turip.data.content.datasource.ContentRemoteDataSource
+import com.on.turip.data.content.place.toDomain
 import com.on.turip.data.content.toDomain
 import com.on.turip.domain.content.Content
 import com.on.turip.domain.content.PagedContentsResult
 import com.on.turip.domain.content.UsersLikeContent
 import com.on.turip.domain.content.repository.ContentRepository
+import com.on.turip.domain.trip.Trip
 import javax.inject.Inject
 
 class DefaultContentRepository @Inject constructor(
@@ -49,5 +51,10 @@ class DefaultContentRepository @Inject constructor(
     override suspend fun loadPopularFavoriteContents(size: Int): TuripResult<List<UsersLikeContent>> =
         contentRemoteDataSource
             .getUsersLikeContents(size)
+            .mapCatching { it.toDomain() }
+
+    override suspend fun loadTripInfo(contentId: Long): TuripResult<Trip> =
+        contentRemoteDataSource
+            .getTrip(contentId)
             .mapCatching { it.toDomain() }
 }
